as-a-function/microsoft/requirements.txt
azure-functions
pdf2image
